[gvf] Move parametric trajectories to trajectories/ folder
