+++ conf/userconf/conf_example_gvf.xml
@@ -4,7 +4,7 @@
    ac_id="1"
    airframe="airframes/UGR/sonic_n1.xml"
    radio="radios/T7C_SBUS.xml"
-   telemetry="telemetry/UGR/gvf_default_fixedwing.xml"
+   telemetry="telemetry/UGR/gvf_fixedwing.xml"
    flight_plan="flight_plans/demo_gvf.xml"
    settings="settings/UGR/fixedwing_basic.xml"
    settings_modules="modules/ahrs_float_dcm.xml modules/electrical.xml modules/gps.xml modules/gps_ublox.xml modules/gps_ubx_ucenter.xml modules/guidance_full_pid_fw.xml modules/gvf_classic.xml~GVF~ modules/gvf_ik.xml~GVF_IK~ modules/gvf_parametric.xml~GVF_PARAMETRIC~ modules/imu_common.xml modules/nav_basic_fw.xml modules/power_switch.xml modules/stabilization_adaptive_fw.xml"
