+++ sw/airborne/modules/guidance/gvf_parametric/nav/nav_parametric_2d_bezier_splines.h
@@ -25,7 +25,7 @@
 extern "C" {
 #endif
 
-#include "modules/guidance/gvf_parametric/gvf_param_traj.h"
+#include "modules/guidance/trajectories/gvf_param_traj.h"
 
 typedef struct {
   float kx;
